fix: restore blog page load function
